refactor generate
